requirements.txt
anyascii==0.3.1
anyio==3.6.2
apturl==0.5.2
asgiref==3.4.1
asn1crypto==0.24.0
beautifulsoup4==4.9.3
blis==0.7.9
bootstrap4==0.1.0
Brlapi==0.6.6
catalogue==2.0.8
certifi==2021.10.8
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
colorama==0.3.7
command-not-found==0.3
confection==0.0.4
contextlib2==21.6.0
contextvars==2.4
cryptography==39.0.0
cupshelpers==1.0
cycler==0.11.0
cymem==2.0.7
dataclasses==0.8
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro-info===0.18ubuntu0.18.04.1
dj-database-url==1.2.0
Django==3.2.13
django-allauth==0.52.0
django-cors-headers==3.10.1
django-crispy-forms==1.13.0
django-filter==21.1
django-heroku==0.3.1
django-modelcluster==5.3
django-taggit==1.5.1
django-treebeard==4.5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.4.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
fastapi==0.83.0
filelock==3.4.1
fontawesomefree==6.3.0
html5lib==1.1
httplib2==0.9.2
idna==3.4
immutables==0.19
importlib-metadata==4.8.3
importlib-resources==5.4.0
Jinja2==3.0.3
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
l18n==2021.3
langcodes==3.3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==2.0.1
matplotlib==3.3.4
murmurhash==1.0.9
netifaces==0.10.4
numpy==1.19.5
oauth==1.0.1
oauthlib==3.2.2
olefile==0.45.1
openpyxl==3.0.10
packaging==21.3
pathy==0.10.1
pbr==5.11.0
pexpect==4.2.1
Pillow==8.4.0
pipenv==2022.1.8
platformdirs==2.4.0
preshed==3.0.8
protobuf==3.0.0
psycopg2==2.7.7
pycairo==1.16.2
pycparser==2.21
pycrypto==2.6.1
pycups==1.9.73
pydantic==1.9.2
PyGObject==3.26.1
PyJWT==2.4.0
pymacaroons==0.13.0
pymongo==4.1.1
PyNaCl==1.1.2
pyparsing==3.0.9
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.7
python-debian==0.1.32
python3-openid==3.2.0
pytz==2022.6
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.27.1
requests-oauthlib==1.3.1
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.16.0
smart-open==6.3.0
sniffio==1.2.0
soupsieve==2.3.2.post1
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.2
srsly==2.4.5
starlette==0.19.1
stevedore==3.5.2
stripe==5.0.0
system-service==0.3
systemd-python==234
tablib==3.1.0
telepath==0.3
testresources==2.0.1
thinc==8.1.7
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.1.1
ubuntu-advantage-tools==27.11.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==20.13.2
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wadllib==1.3.2
wagtail==2.15.6
wasabi==1.1.1
webencodings==0.5.1
whitenoise==5.3.0
Willow==1.4.1
xkit==0.0.0
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
zipp==3.6.0
zope.interface==4.3.2
